bd_sage_site/requirements.txt
APScheduler==3.6.3
attrs==19.3.0
blackduck==0.0.38
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-apscheduler==0.3.0
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.5
pytest==5.3.0
python-dateutil==2.8.1
pytimeparse==1.1.8
pytz==2019.3
requests==2.22.0
requests-mock==1.7.0
six==1.13.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
tzlocal==2.0.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
